requirements.txt
Django==2.0.6
djangorestframework==3.8.2
Unidecode==1.0.22
Pillow==5.1.0
djangorestframework-jwt==1.11.0
